Moved Counter logic into its own component. Also added in gobal css import for tailwind
