+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import { Provider } from 'react-redux';
 import type { AppProps } from 'next/app';
 import { store } from '../app/store';
+import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
